requirements.txt
google-adk
google-genai
pydantic
requests
python-dotenv
